backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2

# AI
openai==1.51.0
anthropic==0.36.0

# File parsing
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5

# Activity monitoring
psutil==6.0.0

# Scheduling
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.27.2
